requirements/requirements-prod.txt
Django==2.0.7
django-model-utils==3.1.2
flake8==3.5.0
psycopg2-binary==2.7.5
python-dateutil==2.7.3
python-decouple==3.1
requests-oauthlib==1.0.0
unipath==1.1
uwsgi==2.0.17.1
django-bootstrap4==0.0.7

# Django==2.0.7
# django-cuser==2017.3.16
# django-jsonfield==1.0.1
# django-model-utils==3.1.2
# flake8==3.5.0
# # mccabe==0.6.1
# nodeenv==1.3.2
# # psycopg2==2.7.5
# psycopg2-binary==2.7.5
# pycodestyle==2.3.1
# pyflakes==1.6.0
# # pygal==2.4.0
# python-dateutil==2.7.3
# pytz==2018.5
# PyYAML==3.13
# six==1.11.0
# # testfixtures==6.2.0
# # toml==0.9.4
# virtualenv==16.0.0
